Added AI bully code; update on dungeon generation
Checked out some Youtube vids about my problem and found a potential solution. I reduced the number of nodes in the room to 4, and have also changed the room to a single cube, but enlarged. I'm going to use that one room to generate the entire dungeon based on what I learned from the video.