created Loot table
Made a loot table that will be used to generate items using a weight system. All items in the game must go here, and each item will have their own weight. Certain dungeon mods will be able to influence an item's weight.

Loot table works by first rolling to see which category of items is accessed. Once that's resolved, the next step is to roll to see which item is generated. For equipment, only equipment types are generated. The actual equipment is generated in Item Manager, complete with mods.