Added AI code for moving CPU Hunter
The CPU will start by looking for a target, and move towards it. Something I'll have to do before I move the hunter is decide which skill they're going to use, because if the hunter has a ranged skill, they can't get too close to their target. What I'll end up doing is as the hunter moves, at each step I need to check the skill range. If the target is in range, I stop moving the CPU and go right into attacking.