working on CPU character targeting while moving
On each step the character takes, they use their chosen skill to see if their target is in range. If they are, they stop moving.

I may change the code so that the parameter for the character is actually the target, and the CPU must move in its range. Once the CPU is standing in range, they can attack.