Added transition to combat
The game transitions to combat by activating Combat game object, and temporarily hiding dungeon layout and Hunter UI. The current problem is that when dungeon is hidden, the characters fall. I might be able to prevent this by deactivating all the characters, but I should keep track of their positions before deactivating the dungeon.