Updated code for MOV cards
Updated the function of all MOV cards so that the attacker can use them to prevent the defender from escaping. If both hunters use an exit card, the attacker should win since they have a natural bonus against escape attempts.

Also, added a reference to the exit point in Dungeon script so I don't have to search the entire dungeon for it when using the Exit card.