Working on Risky Tackle skill; chests look better now
Figured out why the chests weren't displaying correctly. The bottom of the chests can't be angled like the rooms are, they have to be nearly flat. Also reduced the scaling on the chests so they look better.

Added extra coroutines so Risky tackle works like I want it to, and probably future skills.