Changed damage calculations
I felt that the damage was out of control, so now the dice rolls add a percentage to the ATP/DFP/MNP/RST.

The total dice value is converted to a percentage. For example, rolling a 7 increases the attacker's ATP by 70% instead of adding a flat 7 to the ATP. Flat values result in sharper increases than percentages, which means that the damage can potentially wipe out characters faster than desired.