Fixed small errors in combat calculaitons
I realized defensive rolls resulted in the entire DFP value not being used, so made an adjustment there. A side effect to this is that if the defender defends, they get a slight bonus to the calculations since 2 dice are rolled.

Fixed issue where defense boost was not reversed after guarding.