Added code for Traps; added card code
Traps are scriptable objects. Each room can have a reference to a trap.

I added a damage trap but haven't tested it yet.

Need to add code that checks each room for a trap every time a player steps into it.

Added a quick way to reference hunters in the HUD instead of searching for them every time.

Added a bunch of code for the cards so they can actually be used. Trap cards not complete.