Damage trap working; added trap and card IDs
Damage trap works. The game checks the room for a trap every time a character takes a step. Next, I want to add a coroutine that will handle trap evasion/triggering trap.

Added card and Trap IDs to make it easy to grab specific entities. Trying to find a way to do this for items instead of searching by string.