Added Vise trap and DIsable Leg effect
Vise trap works, but might need to adjust the calculations to round down the result.

During testing, I tried fighting the monster but got stuck. I believe this is due to the hunter and monster's Y positions being different. Updated basic attack to prevent trying to match the Y position when moving.