Fixed issue with defender not counterattacking
Defender was not able to move to position to counter attack because I run the same coroutine to counterattack. The problem with that is that there's code after the yield that was interfering with the defender's movement. So I removed that code block and put it into its own coroutine, which I should have done before.