Monster is now removed from game
When monster is defeated, they're removed from the dungeon and the game passes the turn to the next character. However, for some reason the dice UI is on screen and I'm not sure why it's appearing. I assume it has to do with the monster state, which shouldn't exist anymore.